+++ src/communication/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(authenticator)
 
-add_library(communication STATIC Communicator.cpp SSLSocket.cpp MessageProcessor.cpp SchedulerClientMP.cpp SchedulerWorkerMP.cpp)
+add_library(communication STATIC Communicator.cpp SSLSocket.cpp MessageProcessor.cpp)
 target_link_libraries(communication Qt5::Core Qt5::Network)
